# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Este es el proyecto final de la materia espcialización backend 3 para crear, editar, consultar y borrar las entidades de paciente, dentista y turnos.",
        "title": "Proyecto Final Especialización backend 3",
        "contact": {
            "name": "API Support (Natalia Garcia, Federico Bustamante, Damian, Camilo Zuleta)"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http:www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/patients": {
            "post": {
                "description": "Creates a Patient, you must send the fields required to process your request they are name, surname, address, dni, and registration date.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patient"
                ],
                "summary": "Creates a Patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "create Patient",
                        "name": "Patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/patients.Patient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/patients.Patient"
                        }
                    }
                }
            }
        },
        "/patients/{id}": {
            "get": {
                "description": "Gets a Patient by id using the repository principal",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patient"
                ],
                "summary": "Gets a Patient by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/patients.Patient"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a Patient by ID, you must send all of the patient fields to process your request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patient"
                ],
                "summary": "Updates a Patient by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update Patient",
                        "name": "Patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/patients.Patient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/patients.Patient"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a Patient by ID, be careful with this option!",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patient"
                ],
                "summary": "Deletes a Patient by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/patients.Patient"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates a Patient by ID, you may be noticed is not required to send data in all of the fields",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patient"
                ],
                "summary": "Updates a Patient by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update Patient",
                        "name": "Patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/patients.Patient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/patients.Patient"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "patients.Patient": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "dni": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "registration_date": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        }
    }
}